v0.9 Refactored structure of controllers and commands, moved migration addon to a separate module
