+++ CHANGELOG.md
@@ -3,6 +3,11 @@ CHANGELOG for Yii2 STARTER PROJECT TEMPLATE
 
 *should be replaced with real project changelog later*
 
+v0.9
+---------------------
+* Upgrade app folder structure: moved controllers from web to root, moved commands to console, 
+* Moved migration component/views custom configuration to a separate module.
+
 v0.8.6
 ---------------------
 * Fixed unit/functional tests
