--- app/console/views/migrate/dropColumnMigration.php
@@ -1,64 +1,0 @@
-<?php
-/**
- * This view is used by console/controllers/MigrateController.php.
- *
- * The following variables are available in this view:
- */
-/* @var $className string the new migration class name without namespace */
-/* @var $namespace string the new migration class namespace */
-/* @var $table string the name table */
-/* @var $fields array the fields */
-preg_match('/^drop_(.+)_columns?_from_(.+)_table$/', $name, $matches);
-$columns = $matches[1];
-
-if (empty($fields)) {
-	$fields = [
-		[
-			'property' => $columns,
-			'decorators' => 'string()',
-		],
-	];
-}
-
-echo "<?php\n";
-if (!empty($namespace)) {
-    echo "\nnamespace {$namespace};\n";
-}
-?>
-
-use app\console\Migration;
-
-/**
- * Handles dropping <?= $columns ?> from table `<?= $table ?>`.
-<?= $this->render('_foreignTables', [
-    'foreignKeys' => $foreignKeys,
-]) ?>
- */
-class <?= $className ?> extends Migration
-{
-	/**
-	 * @inheritdoc
-	 */
-	public function up()
-	{
-<?= $this->render('_dropColumns', [
-	'table' => $table,
-	'fields' => $fields,
-	'foreignKeys' => $foreignKeys,
-])
-?>
-	}
-
-	/**
-	 * @inheritdoc
-	 */
-	public function down()
-	{
-<?= $this->render('_addColumns', [
-	'table' => $table,
-	'fields' => $fields,
-	'foreignKeys' => $foreignKeys,
-])
-?>
-	}
-}
