+++ app/controllers/AuthController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace app\web\controllers;
+namespace app\controllers;
 
 use Yii;
 use yii\filters\AccessControl;
