+++ app/controllers/Controller.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace app\web\controllers;
+namespace app\controllers;
 
 
 abstract class Controller extends \yii\web\Controller
