+++ app/modules/base/db/Migration.php
@@ -1,5 +1,5 @@
 <?php
-namespace app\console;
+namespace app\modules\base\db;
 
 class Migration extends \yii\db\Migration
 {
@@ -12,7 +12,7 @@ class Migration extends \yii\db\Migration
 	{
 		$tableOptions = null;
 		if ($this->db->driverName === 'mysql') {
-			// http://stackoverflow.com/questions/766809/whats-the-difference-between-utf8-general-ci-and-utf8-unicode-ci
+			// http://stackoverflow.com/questions/766809/whats-the-difference-between-utf8-general-ci-and-utf8-unicode-ci.
 			$tableOptions = 'CHARACTER SET utf8 COLLATE utf8_unicode_ci ENGINE=InnoDB';
 		}
