+++ app/modules/base/views/migrate/_dropColumns.php
