+++ app/modules/base/views/migrate/_dropForeignKeys.php
