+++ app/modules/base/views/migrate/_dropTable.php
