+++ app/modules/base/views/migrate/dropColumnMigration.php
@@ -8,7 +8,17 @@
 /* @var $namespace string the new migration class namespace */
 /* @var $table string the name table */
 /* @var $fields array the fields */
-/* @var $foreignKeys array the foreign keys */
+preg_match('/^drop_(.+)_columns?_from_(.+)_table$/', $name, $matches);
+$columns = $matches[1];
+
+if (empty($fields)) {
+	$fields = [
+		[
+			'property' => $columns,
+			'decorators' => 'string()',
+		],
+	];
+}
 
 echo "<?php\n";
 if (!empty($namespace)) {
@@ -16,12 +26,12 @@ if (!empty($namespace)) {
 }
 ?>
 
-use app\console\Migration;
+use app\modules\base\db\Migration;
 
 /**
- * Handles the creation of table `<?= $table ?>`.
+ * Handles dropping <?= $columns ?> from table `<?= $table ?>`.
 <?= $this->render('_foreignTables', [
-	'foreignKeys' => $foreignKeys,
+    'foreignKeys' => $foreignKeys,
 ]) ?>
  */
 class <?= $className ?> extends Migration
@@ -31,7 +41,7 @@ class <?= $className ?> extends Migration
 	 */
 	public function up()
 	{
-<?= $this->render('_createTable', [
+<?= $this->render('_dropColumns', [
 	'table' => $table,
 	'fields' => $fields,
 	'foreignKeys' => $foreignKeys,
@@ -44,8 +54,9 @@ class <?= $className ?> extends Migration
 	 */
 	public function down()
 	{
-<?= $this->render('_dropTable', [
+<?= $this->render('_addColumns', [
 	'table' => $table,
+	'fields' => $fields,
 	'foreignKeys' => $foreignKeys,
 ])
 ?>
